# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Gin chat documentation",
        "title": "Swagger Example API",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/": {
            "post": {
                "description": "1 min for every request, not authenticated, and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "send token",
                "parameters": [
                    {
                        "description": "Register details",
                        "name": "Register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.RegisterResponse"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Too_Many_Token_Request(7) | Token_Expired_Or_Invalid(2) | We_Don't_Know_What_Happened(8) | MUST_NOT_AUTHENTICATED(1)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Authenticates a user and returns a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "check token",
                "parameters": [
                    {
                        "description": "Login details",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Token"
                        }
                    },
                    "400": {
                        "description": "Object_Not_Found(6) | Token_Expired_Or_Invalid(2) | Name_Field_Required_For_Register(12) | We_Don't_Know_What_Happened(8) | MUST_NOT_AUTHENTICATED(1)",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }

        },
        "/chat/send": {
            "post": {
                "description": "Sends a chat message to a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Send Chat Message",
                "responses": {}
            }
        }
    },
    "definitions": {
        "serializer.LoginRequest": {
            "type": "object",
            "required": [
                "phone_no",
                "token"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone_no": {
                    "type": "string"
                },
                "token": {
                    "type": "integer"
                }
            }
        },
        "serializer.RegisterRequest": {
            "type": "object",
            "required": [
                "phone_no"
            ],
            "properties": {
                "phone_no": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 11
                }
            }
        },
        "serializer.Token": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "utils.RegisterResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "isSignup": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Json Web Token (jwt)": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI"
    }
}